To be more Objective-Cy :p

+++ AAChartKitLib/AAChartConfiger/AAChartView.h
@@ -172,14 +172,14 @@
  */
 - (void)aa_updateChartWithOptions:(NSObject *)options redraw:(BOOL)redraw;
 
-- (void)aa_addPointToChartSeriesWithSeriesElementIndex:(unsigned int)elementIndex
+- (void)aa_addPointToChartSeriesWithSeriesElementIndex:(NSUInteger)elementIndex
                                                options:(NSObject *)options;
 
-- (void)aa_addPointToChartSeriesWithSeriesElementIndex:(unsigned int)elementIndex
+- (void)aa_addPointToChartSeriesWithSeriesElementIndex:(NSUInteger)elementIndex
                                                options:(NSObject *)options
                                                  shift:(BOOL)shift;
 
-- (void)aa_addPointToChartSeriesWithSeriesElementIndex:(unsigned int)elementIndex
+- (void)aa_addPointToChartSeriesWithSeriesElementIndex:(NSUInteger)elementIndex
                                                options:(NSObject *)options
                                                 redraw:(BOOL)redraw
                                                  shift:(BOOL)shift
